Evaluations now use floats

+++ Assets/AI/Evaluation.cs
@@ -1,14 +1,14 @@
 namespace Hikari.AI {
     public struct Evaluation {
-        public int defensiveness;
-        public int offensiveness;
+        public float defensiveness;
+        public float offensiveness;
 
-        public Evaluation(int defensiveness, int offensiveness) {
+        public Evaluation(float defensiveness, float offensiveness) {
             this.defensiveness = defensiveness;
             this.offensiveness = offensiveness;
         }
 
-        public int Sum() => offensiveness + defensiveness; //TODO Sum() should be more appropriate method
+        public float Sum() => offensiveness + defensiveness; //TODO Sum() should be more appropriate method
 
         public override string ToString() {
             return $"{defensiveness} {offensiveness}";
